Make default webengineprofile persistent again.
Qt6 has switched to the global default profile being off-the-record,
which is not the default for qutebrowser.

Restore the previous default object name (although we always point it
elsewhere anyway so that name should appear in any file paths).

ref https://bugreports.qt.io/browse/QTBUG-66068

On Qt 5, .defaultProfile() is still used, to ensure the same behavior.
It's not entirely clear from the Qt 5 source what the differences are
between the two.

(cherry picked from commit b961c3b8201d987c5dcaecff575070c655ec27f1)